Changed wolf and player transform z coordinates to 0.
Changed wolf and player transform z coordinates to 0. Enabled fireball pushForce value. Set all pushForce values to 1.0f through the Unity editor (values written in the code were overwritten by the Unity editor values anyway). The wolf is not pushed back if it is moving, this might be a feature instead of a bug, because it makes sense combat design wise.